Move clusters code to classes
